refactor(bigquery): reorg jobs integration tests
